refactor: bus/route 경로 수정
