+++ src/static/routes.ts
@@ -19,6 +19,7 @@ const ROUTES = {
   BenefitStore: () => '/benefitstore',
   StoreDetail: ({ id, isLink }: ROUTESParams<'id'>) => (isLink ? `/store/${id}` : '/store/:id'),
   Bus: () => '/bus',
+  BusRoute: () => '/bus/route',
   Cafeteria: () => '/cafeteria',
   BoardNotice: () => '/board/notice',
   BoardNoticeDetail: ({ id, isLink }: ROUTESParams<'id'>) => (isLink ? `/board/notice/${id}` : '/board/notice/:id'),
